+++ ivetl/tools/normalize/highwire_metadata.go
@@ -43,10 +43,9 @@ func normalizeHighwireMetadata(ctx context.Context, session *gocql.Session, meta
 							if doiType != MATCH_NO_DOI {
 								s = doi
 							} else {
-								// DISABLING THIS FOR NOW, PROBABLY
-								// WE JUST WANT HUMAN INTERVENTION
-								// invalidDOI = true
-								log.Printf("invalid DOI for highwire_metadata row %v: %s", pk, s)
+								log.Printf("invalid DOI for %s.%s row %v: %s",
+									col.Keyspace, col.Table, pk, s)
+								invalidDOI = true
 							}
 						}
 
@@ -85,8 +84,7 @@ func normalizeHighwireMetadata(ctx context.Context, session *gocql.Session, meta
 			return fmt.Errorf("error initializing update statement: %w", err)
 		}
 
-		// if invalidDOI is true, delete rows in this set and return
-		if invalidDOI {
+		if invalidDOI && deleteInvalidDOI {
 			for i := range set {
 				col, val := delete_bind(set[i], nil)
 				dq := session.Query(delete_stmt, val...)
