+++ ivetl/tools/normalize/published_article_values.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"log"
 	"reflect"
 	"strings"
 
@@ -41,8 +42,9 @@ func normalizePublishedArticleValues(ctx context.Context, session *gocql.Session
 							if doiType != MATCH_NO_DOI {
 								s = doi
 							} else {
-								// don't actually flag it
-								// invalidDOI = true
+								log.Printf("invalid DOI for %s.%s row %v: %s",
+									col.Keyspace, col.Table, pk, s)
+								invalidDOI = true
 							}
 						}
 
@@ -77,8 +79,7 @@ func normalizePublishedArticleValues(ctx context.Context, session *gocql.Session
 			return fmt.Errorf("error initializing update statement: %w", err)
 		}
 
-		// if invalidDOI is true, delete rows in this set and return
-		if invalidDOI {
+		if invalidDOI && deleteInvalidDOI {
 			for i := range set {
 				col, val := delete_bind(set[i], nil)
 				dq := session.Query(delete_stmt, val...)
